+++ Source/Rhetos.Utilities/AppSecurityOptions.cs
@@ -22,7 +22,6 @@ namespace Rhetos.Utilities
     [Options("Rhetos:AppSecurity")]
     public class AppSecurityOptions
     {
-        public bool BuiltinAdminOverride { get; set; } = false;
         public bool LookupClientHostname { get; set; } = false;
         public string AllClaimsForUsers { get; set; } = "";
         public bool AllClaimsForAnonymous { get; set; } = false;